Fix: no need for extra module and only print if properly installed anyway
