+++ mdsobjects/python/__init__.py
@@ -68,10 +68,29 @@ try:
     del _mvers
 except:
     __version__='Unknown'
-
-_verscheck=_mimport('version_check')
-_verscheck.version_check(__version__)
-del _verscheck
+else:
+    try:
+        import sys,ctypes as _C
+        class MDSplusVersionInfo(_C.Structure):
+            _fields_= [("MAJOR",_C.c_char_p),
+                       ("MINOR",_C.c_char_p),
+                       ("RELEASE",_C.c_char_p),
+                       ("BRANCH",_C.c_char_p),
+                       ("RELEASE_TAG",_C.c_char_p),
+                       ("COMMIT",_C.c_char_p),
+                       ("DATE",_C.c_char_p),
+                       ("MDSVERSION",_C.c_char_p)]
+        mdsshr=_ver.load_library('MdsShr')
+        _info=_C.cast(mdsshr.MDSplusVersion,_C.POINTER(MDSplusVersionInfo)).contents
+        if _info.BRANCH in ('stable','alpha') and _info.MDSVERSION != __version__:
+            sys.stderr.write("""Warning:
+  The MDSplus python module version (%s) does not match
+  the version of the installed MDSplus libraries (%s).
+  Upgrade the module using the mdsplus/mdsobjects/python directory of the
+  MDSplus installation.
+""" % (version, _info.MDSVERSION ))
+    except:
+        pass
 
 def load(gbls=globals()):
     def loadmod_full(name,gbls):
